CentralizedTelescope: Rename wrapper
